Allow active powerups to reset on powerup pickup
This rule follows the same logic as Mario's star power-up, as in players should not be forced to wait out their power-up if they want to pick-up the same power-up again.

This is only applied to co-op/single-player gameplay, multiplayer will still follow the same original design (do not reset powerup time).